Dedup Windows CentralProcessor frequency queries; fix JNA numbering
Hoist queryCurrentFreq, mapPercentToFreqs, and queryMaxFreq into the shared
WindowsCentralProcessor base, with abstract hooks for the backend-specific
native calls (queryProcessorPerformanceCounters, queryFrequencyCounters,
queryNTPower) and a native-free isWindows7OrGreater() version check.

The JNA mapPercentToFreqs, added in #3233, indexed the parallel percent list
by computed CPU number and defaulted unmapped NUMA instances to CPU 0, so on
multi-processor-group / NUMA systems it assigned current frequencies to the
wrong logical processors. Unify on the positional numbering already used by
processTickData (the #1373-hardened usage path) and the FFM frequency code,
fixing the JNA path.
